# Add support for DeepseekV2ForCausalLM