@fairydreaming 2024-05-27
> I think these [issues about YaRN](https://github.com/ggerganov/ggml/issues/835) are not addressed yet.

@foldl While I agree that there may be some issues with YaRN implementation for models using NeoX-style RoPE (at least by looking at the code), I think DeepSeek-V2 is not affected by them, as it falls into preceding conditional block (`if(!is_neox)`):

https://github.com/ggerganov/ggml/blob/0cbb7c0e053f5419cfbebb46fbf4d4ed60182cf5/src/ggml.c#L14059C19-L14059C39