@foldl 2024-05-27
@fairydreaming Got it. Your PR uses RoPE type `0`.