@bartowski1182 2024-05-28
Do we expect this to not work on CUDA?

Trying to make imatrix for https://huggingface.co/deepseek-ai/DeepSeek-V2-Lite-Chat

Results in: 

GGML_ASSERT: ggml-cuda/concat.cu:107: ggml_is_contiguous(src0)