@fairydreaming 2024-05-29
> @fairydreaming Could you give #7617 a try and see if it works correctly?

@ggerganov  Sure, seems to work OK both on CPU and on GPU (CUDA).

> I ended up implementing your patch because the idea that I had is not going to work. The partial RoPE requires that the rotated dimensions are at the beginning of the head, but here they are at the end. There also seem to be other limitations, so I guess the current `view + cont` implementation should be ok

Roger that, I checked the performance difference on the CPU caused by added `ggml_cont` operations and it's negligible (around 2%):

without ggml_cont:

```
llama_print_timings: prompt eval time =     106.20 ms /    17 tokens (    6.25 ms per token,   160.07 tokens per second)
llama_print_timings:        eval time =   12712.43 ms /   472 runs   (   26.93 ms per token,    37.13 tokens per second)
```

with ggml_cont:

```
llama_print_timings: prompt eval time =     108.40 ms /    17 tokens (    6.38 ms per token,   156.83 tokens per second)
llama_print_timings:        eval time =   12875.21 ms /   472 runs   (   27.28 ms per token,    36.66 tokens per second)
```

> The CUDA kernels could be extended to support non-contiguous RoPE and norm, but I'm not sure if this is going to be worth it - I guess it might even hurt the performance. So instead, I added asserts to prevent from using non-contiguous data with these ops

If needed I guess there could be separate implementations for contiguous and non-contiguous tensors. However, it looks like a lot of work and the efficiency gains won't be large (or like you said on CUDA the performance may be even worse), so IMHO it can stay this way.